+++ packages/core/src/step/get-workflow-metadata.ts
@@ -1,4 +1,4 @@
-import * as Logger from '../prettylogger.js';
+import { Ansi } from '@workflow/errors';
 import type { WorkflowMetadata } from '../workflow/get-workflow-metadata.js';
 import { contextStorage } from './context-storage.js';
 
@@ -11,10 +11,10 @@ export function getWorkflowMetadata(): WorkflowMetadata {
   const ctx = contextStorage.getStore();
   if (!ctx) {
     throw new Error(
-      Logger.frame(
+      Ansi.frame(
         '`getWorkflowMetadata()` can only be called inside a workflow or step function',
         [
-          Logger.help([
+          Ansi.help([
             'This function comes from Workflow DevKit, and requires to be used as a part of a workflow or a step,',
             'As it has no meaning outside of the workflow context.',
             'Read more: https://useworkflow.dev/docs/api-reference/workflow/get-workflow-metadata',
