+++ packages/world-local/src/config.ts
@@ -1,6 +1,5 @@
+import { Ansi } from '@workflow/errors';
 import { getWorkflowPort } from '@workflow/utils/get-port';
-import { frame } from './frame.js';
-import * as Logger from './logger.js';
 import { once } from './util.js';
 
 const getDataDirFromEnv = () => {
@@ -52,19 +51,16 @@ export async function resolveBaseUrl(config: Partial<Config>): Promise<string> {
   }
 
   throw new Error(
-    frame({
-      text: `Unable to resolve base URL for workflow queue`,
-      contents: [
-        'The local world works by making HTTP calls to the .well-known/workflow endpoints[1].\n' +
-          'Therefore, it needs to have a base URL to connect to the local server.',
-        Logger.note('we tried inferring the running port but failed.'),
-        Logger.help([
-          `fix by setting one of the following environment variables:`,
-          `• ${Logger.code('PORT')} to use ${Logger.code('http://localhost:PORT')}`,
-          `• ${Logger.code('WORKFLOW_LOCAL_BASE_URL')} as a full URL`,
-        ]),
-        'Read more about .well-known endpoints: https://useworkflow.dev/docs/how-it-works/framework-integrations#understanding-the-endpoints',
-      ],
-    })
+    Ansi.frame(`Unable to resolve base URL for workflow queue`, [
+      'The local world works by making HTTP calls to the .well-known/workflow endpoints[1].\n' +
+        'Therefore, it needs to have a base URL to connect to the local server.',
+      Ansi.note('we tried inferring the running port but failed.'),
+      Ansi.help([
+        `fix by setting one of the following environment variables:`,
+        `• ${Ansi.code('PORT')} to use ${Ansi.code('http://localhost:PORT')}`,
+        `• ${Ansi.code('WORKFLOW_LOCAL_BASE_URL')} as a full URL`,
+      ]),
+      'Read more about .well-known endpoints: https://useworkflow.dev/docs/how-it-works/framework-integrations#understanding-the-endpoints',
+    ])
   );
 }
